// src/app/core/api/models/paginate-expense-detail.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ExpenseDetail } from '../models/expense-detail';
export interface PaginateExpenseDetail {
  content?: Array<ExpenseDetail>;
  currentSize?: number;
  nextPageIndex?: number;
  pageIndex?: number;
  pageSize?: number;
  prevPageIndex?: number;
  totalPages?: number;
  totalSize?: number;
}